backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.24.0.post1
SQLAlchemy==2.0.23
pydantic==1.10.13
python-dotenv==1.0.0
